refactor: rename the base module
